+++ CHANGELOG.md
@@ -10,13 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- run postsetup.sh at the end of setup.sh (#11)
 - optional test file argument for builder.py (#12)
+- type hints for all API functions (#13)
 
 ### Changed
 
 - `assert_not_imported` can also take a string
 - `redirect_stdin` also wraps `builtins.input`
-- run postsetup.sh at the end of setup.sh (#11)
 
 
 ## [1.4.0] - 2025-10-09